requirements.txt
streamlit
pandas
XlsxWriter
openai
altair
python-docx
fpdf
altair_saver
vl-convert-python
langdetect
supabase
pdf2image
pillow
tiktoken
PyPDF2
